Rename models.photon to models.physical
