+++ sunkit_spex/__init__.py
@@ -6,4 +6,4 @@ from .version import __version__
 __all__ = []
 from . import extern
 from .legacy import fitting_legacy
-from .models.photon.thermal import io, thermal
+from .models.physical import io, thermal
