+++ sunkit_spex/models/physical/tests/test_thermal.py
